Rename FidoClient to FidoController
